requirements.txt
astroid==2.1.0
colorama==0.4.1
Django==2.1.7
django-crispy-forms==1.7.2
isort==4.3.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pylint==2.2.2
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytz==2018.9
six==1.12.0
wrapt==1.11.1
